modify rules pattern
