+++ server/engine/src/test/java/org/eclipse/lsp/cobol/usecases/TestCICSEndBrowse.java
@@ -25,23 +25,12 @@ import org.junit.jupiter.api.Test;
 import java.util.Map;
 
 /**
- * Test CICS ENDBR & ENDBROWSE  commands. Documentation link: <a
+ * Test CICS  ENDBROWSE  commands. Documentation link: <a
  * href="https://www.ibm.com/docs/en/cics-ts/6.x?topic=summary-endbrowse-activity">ENDBROWSE Command</a>
- * href="https://www.ibm.com/docs/en/cics-ts/6.x?topic=summary-endbr">ENDBR Command</a>
  *
- * <p>This class tests all variations of the ENDBR & ENDBROWSE command found in the link above.
+ * <p>This class tests all variations of the ENDBROWSE command found in the link above.
  */
 public class TestCICSEndBrowse {
-  //  ENDBR
-  private static final String ENDBR_FILE_VALID = "ENDBR  FILE({$varFour})";
-  private static final String ENDBR_INVALID = "{ENDBR|error1}";
-  private static final String ENDBR_DATASET_VALID = "ENDBR  DATASET({$varFour})";
-  private static final String ENDBR_FILE_DATASET_INVALID = "ENDBR  {FILE|error1}({$varFour}) {DATASET|error2}({$varFour})";
-  private static final String ENDBR_FILE_REQID_VALID = "ENDBR FILE({$varFour}) REQID({$varOne})";
-  private static final String ENDBR_FILE_SYSID_VALID = "ENDBR FILE({$varFour}) SYSID({$varOne})";
-  private static final String ENDBR_FILE_REQID_SYSID_VALID = "ENDBR FILE({$varFour}) REQID({$varOne}) SYSID({$varFive})";
-
-  //  ENDBROWSE
   private static final String ENDBROWSE_ACTIVITY_VALID = "ENDBROWSE ACTIVITY BROWSETOKEN({$varOne})";
   private static final String ENDBROWSE_ACTIVITY_INVALID = "ENDBROWSE ACTIVITY {ACTIVITY|error1} BROWSETOKEN({$varOne})";
   private static final String ENDBROWSE_ACTIVITY_CONTAINER_INVALID = "ENDBROWSE {ACTIVITY|error1} {CONTAINER|error2} BROWSETOKEN({$varOne}) RETCODE({$varOne})";
@@ -51,63 +40,7 @@ public class TestCICSEndBrowse {
   private static final String ENDBROWSE_TIMER_VALID = "ENDBROWSE TIMER BROWSETOKEN({$varOne})";
   private static final String ENDBROWSE_PROCESS_VALID = "ENDBROWSE PROCESS BROWSETOKEN({$varOne})";
 
-  @Test
-  void testEndbrFileValid() {
-    CICSTestUtils.noErrorTest(ENDBR_FILE_VALID);
-  }
-
-  @Test
-  void testEndbrInvalid() {
-    CICSTestUtils.errorTest(ENDBR_INVALID,
-            ImmutableMap.of(
-                    "error1",
-                    new Diagnostic(
-                            new Range(),
-                            "Exactly one option required, none provided: FILE or DATASET",
-                            DiagnosticSeverity.Error,
-                            ErrorSource.PARSING.getText())));
-    }
-
-  @Test
-  void testEndbrDatasetValid() {
-    CICSTestUtils.noErrorTest(ENDBR_DATASET_VALID);
-  }
-
-  @Test
-  void testEndbrFileDatasetInvalid() {
-    Map<String, Diagnostic> expectedDiagnostics =
-            ImmutableMap.of(
-                    "error1",
-                    new Diagnostic(
-                            new Range(),
-                            "Exactly one option required, options are mutually exclusive: FILE or DATASET",
-                            DiagnosticSeverity.Error,
-                            ErrorSource.PARSING.getText()),
-                    "error2",
-                    new Diagnostic(
-                            new Range(),
-                            "Exactly one option required, options are mutually exclusive: FILE or DATASET",
-                            DiagnosticSeverity.Error,
-                            ErrorSource.PARSING.getText()));
-    CICSTestUtils.errorTest(ENDBR_FILE_DATASET_INVALID, expectedDiagnostics);
-  }
-  @Test
-  void testEndbrFileReqidValid() {
-    CICSTestUtils.noErrorTest(ENDBR_FILE_REQID_VALID);
-  }
-
-  @Test
-  void testEndbrFileSysidValid() {
-    CICSTestUtils.noErrorTest(ENDBR_FILE_SYSID_VALID);
-  }
-
-  @Test
-  void testEndbrFileReqidSysidValid() {
-    CICSTestUtils.noErrorTest(ENDBR_FILE_REQID_SYSID_VALID);
-  }
-
-
-  @Test
+    @Test
   void testEndbrowseActivityValid() {
     CICSTestUtils.noErrorTest(ENDBROWSE_ACTIVITY_VALID);
   }
